Add metadata to printable lesson: closes #151
